Changed Pallets to decimal, to allow half pallets. Changed products grid to allow inserting OPTIONAL decimal places

+++ GestionePosizioni/Models/DocumentTypes/ProductDetails.cs
@@ -4,7 +4,7 @@ namespace Models.DocumentTypes
     {
         public int? ProductId { get; set; }
         public string Description { get; set; }
-        public int Pallets { get; set; }
+        public decimal Pallets { get; set; }
         public int Packages { get; set; }
         public decimal GrossWeight { get; set; }
         public decimal NetWeight { get; set; }
